requirements.txt
altgraph==0.17.4
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
decorator==4.4.2
dek==1.3.0
editor==1.4.0
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
joblib==1.3.2
lazy_loader==0.3
moviepy==1.0.3
networkx==3.2
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pefile==2023.2.7
Pillow==9.5.0
pip==23.3.1
proglog==0.1.10
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
runs==1.1.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
setuptools==68.2.2
sklearn==0.0.post10
threadpoolctl==3.2.0
tifffile==2023.9.26
tqdm==4.66.1
urllib3==2.0.7
xmod==1.6.0
